Move from decimal amount to long amount (in cents)

+++ LotteryGame/WalletService/Core/WalletService.Core/Contracts/IFundsOperations.cs
@@ -4,7 +4,7 @@
 
     public interface IFundsOperations
     {
-        Task<ResponseDto<BaseDto>> Reserve(int playerId, decimal amount, string ticketId);
+        Task<ResponseDto<BaseDto>> Reserve(int playerId, long amount, string ticketId);
 
         Task<ResponseDto> Capture(int reservationId);
